+++ libs/core/avatar/avatar.component.ts
@@ -24,7 +24,7 @@ import {
     applyCssClass,
     getRandomColorAccent
 } from '@fundamental-ngx/cdk/utils';
-import { IconComponent } from '@fundamental-ngx/core/icon';
+import { FD_DEFAULT_ICON_FONT_FAMILY, IconComponent } from '@fundamental-ngx/core/icon';
 import { AvatarIconPipe } from './avatar-icon.pipe';
 import { AvatarValueStates } from './avatar-value-states.type';
 import { FD_AVATAR_COMPONENT } from './tokens';
@@ -89,7 +89,7 @@ export class AvatarComponent implements OnChanges, OnInit, CssClassBuilder, OnCh
 
     /** Font family of the icon. */
     @Input()
-    font: IconComponent['font'] = 'SAP-icons';
+    font: IconComponent['font'] = FD_DEFAULT_ICON_FONT_FAMILY;
 
     /** The glyph name. */
     @Input() glyph: Nullable<string> = null;