+++ libs/docs/core/splitter/examples/default/splitter-default-example.component.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 import { SplitterModule, SplitterPaneResizeEvent } from '@fundamental-ngx/core/splitter';
 
 @Component({
-    selector: 'fd-slider-default-example',
+    selector: 'fd-splitter-default-example',
     templateUrl: './splitter-default-example.component.html',
     standalone: true,
     imports: [SplitterModule]