+++ libs/cdk/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.48.0-rc.21](https://github.com/SAP/fundamental-ngx/compare/v0.48.0-rc.20...v0.48.0-rc.21) (2023-12-25)
+
+**Note:** Version bump only for package @fundamental-ngx/cdk
+
+
+
+
+
+# [0.48.0-rc.20](https://github.com/SAP/fundamental-ngx/compare/v0.48.0-rc.19...v0.48.0-rc.20) (2023-12-25)
+
+**Note:** Version bump only for package @fundamental-ngx/cdk
+
+
+
+
+
+# [0.48.0-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.48.0-rc.18...v0.48.0-rc.19) (2023-12-21)
+
+**Note:** Version bump only for package @fundamental-ngx/cdk
+
+
+
+
+
+# [0.48.0-rc.18](https://github.com/SAP/fundamental-ngx/compare/v0.48.0-rc.17...v0.48.0-rc.18) (2023-12-18)
+
+**Note:** Version bump only for package @fundamental-ngx/cdk
+
+
+
+
+
 # [0.48.0-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.48.0-rc.16...v0.48.0-rc.17) (2023-12-15)
 
 **Note:** Version bump only for package @fundamental-ngx/cdk